requirement.txt
altgraph==0.17.5
bcrypt==5.0.0
charset-normalizer==3.4.7
et_xmlfile==2.0.0
macholib==1.16.4
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
python-dateutil==2.9.0.post0
reportlab==4.4.10
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
